+++ app/Http/Controllers/ContaController.php
@@ -10,7 +10,7 @@ class ContaController extends Controller
 {
     public function index()
     {
-      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin && Auth()->user()->email != "admin@test.com"){
         $contas = Conta::all();
         return view('conta.list', compact('contas'));
       }else{
@@ -20,7 +20,7 @@ class ContaController extends Controller
 
     public function create()
     {
-      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin && Auth()->user()->email != "admin@test.com"){
         $conta = new Conta;
         return view('conta.add', compact('conta'));
       }else{
@@ -30,7 +30,7 @@ class ContaController extends Controller
 
     public function store(StoreContaRequest $contaRequest)
     {
-      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin && Auth()->user()->email != "admin@test.com"){
       $fields = $contaRequest->validated();
         $conta = new Conta;
         $conta->fill($fields);
@@ -43,7 +43,7 @@ class ContaController extends Controller
 
     public function show(Conta $conta)
     {
-      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin && Auth()->user()->email != "admin@test.com"){
         return view('conta.show', compact('conta'));
       }else{
           abort(403);
@@ -52,7 +52,7 @@ class ContaController extends Controller
 
     public function edit(Conta $conta)
     {
-      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin && Auth()->user()->email != "admin@test.com"){
         return view('conta.edit', compact('conta'));
       }else{
           abort(403);
@@ -61,7 +61,7 @@ class ContaController extends Controller
 
     public function update(UpdateContaRequest $contaRequest, Conta $conta)
     {
-      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin && Auth()->user()->email != "admin@test.com"){
         $fields = $contaRequest->validated();
         $conta->fill($fields);
         $conta->updated_at = time();
@@ -74,7 +74,7 @@ class ContaController extends Controller
 
     public function destroy(Conta $conta)
     {
-      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+      if (Auth()->user()->tipo == "admin" && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin && Auth()->user()->email != "admin@test.com"){
         $conta->delete();
         return redirect()->route('conta.index')->with('success', 'Conta bancária eliminada com sucesso.');
       }else{
